// frontend/src/app/models/report/reportDepartmentIncomeData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ReportDepartmentIncomeData {
    costShareCenterAmount?: number;
    costShareCenterPercent?: number;
    costShareDepAmount?: number;
    costShareDepPercent?: number;
    costShareGlobalAmount?: number;
    costShareGlobalPercent?: number;
    depId?: number;
    depNameShortEn?: string;
    depNameShortTh?: string;
    first?: number;
    mode?: string;
    remainAmount?: number;
    resultDate?: string;
    resultDateList?: string[];
    rowNum?: number;
    size?: number;
}